requirements.txt
# requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
schedule>=1.1.0
jmcomic>=1.0.0  # 根据实际版本调整